TryCreateEpubAsyncをサービスとして切り出す
