+++ Epub/KoeBook.Epub/Models/EpubDocument.cs
@@ -6,15 +6,6 @@ namespace KoeBook.Epub.Models;
 
 public class EpubDocument(string title, string author, string coverFilePath, Guid id)
 {
-    readonly string _containerXml = """
-     <?xml version="1.0" encoding="UTF-8"?>
-     <container version="1.0" xmlns="urn:oasis:names:tc:opendocument:xmlns:container">
-         <rootfiles>
-             <rootfile full-path="OEBPS/book.opf" media-type="application/oebps-package+xml" />
-         </rootfiles>
-     </container>
-     """;
-
     public string Title { get; set; } = title;
     public string Author { get; set; } = author;
 
@@ -214,97 +205,12 @@ public class EpubDocument(string title, string author, string coverFilePath, Gui
         return builder.ToString();
     }
 
-    public async Task<bool> TryCreateEpubAsync(string tmpDirectory, string name, CancellationToken ct)
-    {
-        if (!File.Exists(CoverFilePath))
-        {
-            throw new FileNotFoundException("指定されたカバーファイルが存在しません", CoverFilePath);
-        }
-        try
-        {
-            using var fs = File.Create(Path.Combine(tmpDirectory, $"{name}.epub"));
-            using var archive = new ZipArchive(fs, ZipArchiveMode.Create);
-
-            var mimeTypeEntry = archive.CreateEntry("mimetype", CompressionLevel.NoCompression);
-            using (var mimeTypeStream = new StreamWriter(mimeTypeEntry.Open()))
-            {
-                await mimeTypeStream.WriteAsync("application/epub+zip").ConfigureAwait(false);
-                await mimeTypeStream.FlushAsync(ct).ConfigureAwait(false);
-            }
-
-            var containerEntry = archive.CreateEntry("META-INF/container.xml");
-            using (var containerStream = new StreamWriter(containerEntry.Open()))
-            {
-                await containerStream.WriteLineAsync(_containerXml).ConfigureAwait(false);
-                await containerStream.FlushAsync(ct).ConfigureAwait(false);
-            }
-
-            archive.CreateEntryFromFile(CoverFilePath, $"OEBPS/{Path.GetFileName(CoverFilePath)}");
-
-            var cssEntry = archive.CreateEntry("OEBPS/style.css");
-            using (var cssStream = new StreamWriter(cssEntry.Open()))
-            {
-                await cssStream.WriteLineAsync(CreateCssText()).ConfigureAwait(false);
-                await cssStream.FlushAsync(ct).ConfigureAwait(false);
-            }
-
-            var navEntry = archive.CreateEntry("OEBPS/nav.xhtml");
-            using (var navStream = new StreamWriter(navEntry.Open()))
-            {
-                await navStream.WriteLineAsync(CreateNavXhtml()).ConfigureAwait(false);
-                await navStream.FlushAsync(ct).ConfigureAwait(false);
-            }
-
-            var opfEntry = archive.CreateEntry("OEBPS/book.opf");
-            using (var opfStream = new StreamWriter(opfEntry.Open()))
-            {
-                await opfStream.WriteLineAsync(CreateOpf()).ConfigureAwait(false);
-                await opfStream.FlushAsync(ct).ConfigureAwait(false);
-            }
-
-            for (var i = 0; i < Chapters.Count; i++)
-            {
-                for (var j = 0; j < Chapters[i].Sections.Count; j++)
-                {
-                    var sectionXhtmlEntry = archive.CreateEntry($"OEBPS/{Chapters[i].Sections[j].Id}.xhtml");
-                    using (var sectionXhtmlStream = new StreamWriter(sectionXhtmlEntry.Open()))
-                    {
-                        await sectionXhtmlStream.WriteLineAsync(Chapters[i].Sections[j].CreateSectionXhtml()).ConfigureAwait(false);
-                        await sectionXhtmlStream.FlushAsync(ct).ConfigureAwait(false);
-                    }
-                    var sectionSmilEntry = archive.CreateEntry($"OEBPS/{Chapters[i].Sections[j].Id}_audio.smil");
-                    using (var sectionSmilStream = new StreamWriter(sectionSmilEntry.Open()))
-                    {
-                        await sectionSmilStream.WriteLineAsync(Chapters[i].Sections[j].CreateSectionSmil()).ConfigureAwait(false);
-                        await sectionSmilStream.FlushAsync(ct).ConfigureAwait(false);
-                    }
-                    for (var k = 0; k < Chapters[i].Sections[j].Elements.Count; k++)
-                    {
-                        var element = Chapters[i].Sections[j].Elements[k];
-                        if (element is Paragraph para && para.Audio != null)
-                        {
-                            var audioEntry = archive.CreateEntry($"OEBPS/{Chapters[i].Sections[j].Id}_p{k}.mp3");
-                            using var audioStream = para.Audio.GetStream();
-                            using var audioEntryStream = audioEntry.Open();
-                            await audioStream.CopyToAsync(audioEntryStream, ct).ConfigureAwait(false);
-                            await audioEntryStream.FlushAsync(ct).ConfigureAwait(false);
-                        }
-                        else if (element is Picture pic && File.Exists(pic.PictureFilePath))
-                        {
-                            archive.CreateEntryFromFile(pic.PictureFilePath, $"OEBPS/{Chapters[i].Sections[j].Id}_p{k}{Path.GetExtension(pic.PictureFilePath)}");
-                        }
-                    }
-                }
-            }
-            return true;
-        }
-        catch (OperationCanceledException)
-        {
-            throw;
-        }
-        catch
-        {
-            return false;
-        }
-    }
+    public string CreateContainerXml() => """
+     <?xml version="1.0" encoding="UTF-8"?>
+     <container version="1.0" xmlns="urn:oasis:names:tc:opendocument:xmlns:container">
+         <rootfiles>
+             <rootfile full-path="OEBPS/book.opf" media-type="application/oebps-package+xml" />
+         </rootfiles>
+     </container>
+     """;
 }
